Moved remaining functionality into GenericIterativeSolve class and standardized all inputs to be in double precision with explicity specification of type now required

+++ test/test_GMRES/test_GMRES_solve_hlf.cpp
@@ -25,9 +25,9 @@ class GMRESSolveHalfTest: public TestBase {
 TEST_F(GMRESSolveHalfTest, SolveConvDiff64) {
 
     constexpr int n(64);
-    Matrix<half, n, n> A(read_matrix_csv<half>(solve_matrix_dir + "conv_diff_64_A.csv"));
-    Matrix<half, n, 1> b(read_matrix_csv<half>(solve_matrix_dir + "conv_diff_64_b.csv"));
-    GMRESSolve<half> gmres_solve_h(A, b, static_cast<half>(u_hlf), n, conv_tol_hlf);
+    Matrix<double, n, n> A(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_A.csv"));
+    Matrix<double, n, 1> b(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_b.csv"));
+    GMRESSolve<half> gmres_solve_h(A, b, u_hlf, n, conv_tol_hlf);
 
     gmres_solve_h.solve();
     if (show_plots) { gmres_solve_h.view_relres_plot("log"); }
@@ -40,10 +40,10 @@ TEST_F(GMRESSolveHalfTest, SolveConvDiff64) {
 TEST_F(GMRESSolveHalfTest, SolveConvDiff256) {
 
     constexpr int n(256);
-    Matrix<half, Dynamic, Dynamic> A(read_matrix_csv<half>(solve_matrix_dir + "conv_diff_256_A.csv"));
-    Matrix<half, n, 1> b(read_matrix_csv<half>(solve_matrix_dir + "conv_diff_256_b.csv"));
+    Matrix<double, Dynamic, Dynamic> A(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_256_A.csv"));
+    Matrix<double, n, 1> b(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_256_b.csv"));
     GMRESSolve<half> gmres_solve_h(
-        A, b, static_cast<half>(u_hlf), n, large_matrix_error_mod_stag*conv_tol_hlf
+        A, b, u_hlf, n, large_matrix_error_mod_stag*conv_tol_hlf
     );
 
     gmres_solve_h.solve();
@@ -57,10 +57,10 @@ TEST_F(GMRESSolveHalfTest, SolveConvDiff256) {
 TEST_F(GMRESSolveHalfTest, SolveConvDiff1024_LONGRUNTIME) {
 
     constexpr int n(1024);
-    Matrix<half, Dynamic, Dynamic> A(read_matrix_csv<half>(solve_matrix_dir + "conv_diff_1024_A.csv"));
-    Matrix<half, n, 1> b(read_matrix_csv<half>(solve_matrix_dir + "conv_diff_1024_b.csv"));
+    Matrix<double, Dynamic, Dynamic> A(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_1024_A.csv"));
+    Matrix<double, n, 1> b(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_1024_b.csv"));
     GMRESSolve<half> gmres_solve_h(
-        A, b, static_cast<half>(u_hlf), n, large_matrix_error_mod_stag*conv_tol_hlf
+        A, b, u_hlf, n, large_matrix_error_mod_stag*conv_tol_hlf
     );
 
     gmres_solve_h.solve();
@@ -74,9 +74,9 @@ TEST_F(GMRESSolveHalfTest, SolveConvDiff1024_LONGRUNTIME) {
 TEST_F(GMRESSolveHalfTest, SolveRand20) {
 
     constexpr int n(20);
-    Matrix<half, n, n> A(read_matrix_csv<half>(solve_matrix_dir + "A_20_rand.csv"));
-    Matrix<half, n, 1> b(read_matrix_csv<half>(solve_matrix_dir + "b_20_rand.csv"));
-    GMRESSolve<half> gmres_solve_h(A, b, static_cast<half>(u_hlf), n, conv_tol_hlf);
+    Matrix<double, n, n> A(read_matrix_csv<double>(solve_matrix_dir + "A_20_rand.csv"));
+    Matrix<double, n, 1> b(read_matrix_csv<double>(solve_matrix_dir + "b_20_rand.csv"));
+    GMRESSolve<half> gmres_solve_h(A, b, u_hlf, n, conv_tol_hlf);
 
     gmres_solve_h.solve();
     if (show_plots) { gmres_solve_h.view_relres_plot("log"); }
@@ -89,9 +89,9 @@ TEST_F(GMRESSolveHalfTest, SolveRand20) {
 TEST_F(GMRESSolveHalfTest, Solve3Eigs) {
 
     constexpr int n(25);
-    Matrix<half, n, n> A(read_matrix_csv<half>(solve_matrix_dir + "A_25_3eigs.csv"));
-    Matrix<half, n, 1> b(read_matrix_csv<half>(solve_matrix_dir + "b_25_3eigs.csv"));
-    GMRESSolve<half> gmres_solve_h(A, b, static_cast<half>(u_hlf), n, conv_tol_hlf);
+    Matrix<double, n, n> A(read_matrix_csv<double>(solve_matrix_dir + "A_25_3eigs.csv"));
+    Matrix<double, n, 1> b(read_matrix_csv<double>(solve_matrix_dir + "b_25_3eigs.csv"));
+    GMRESSolve<half> gmres_solve_h(A, b, u_hlf, n, conv_tol_hlf);
 
     gmres_solve_h.solve();
     if (show_plots) { gmres_solve_h.view_relres_plot("log"); }
@@ -105,11 +105,11 @@ TEST_F(GMRESSolveHalfTest, Solve3Eigs) {
 TEST_F(GMRESSolveHalfTest, DivergeBeyondHalfCapabilities) {
 
     constexpr int n(64);
-    Matrix<half, n, n> A(read_matrix_csv<half>(solve_matrix_dir + "conv_diff_64_A.csv"));
-    Matrix<half, n, 1> b(read_matrix_csv<half>(solve_matrix_dir + "conv_diff_64_b.csv"));
+    Matrix<double, n, n> A(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_A.csv"));
+    Matrix<double, n, 1> b(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_b.csv"));
 
-    // Check convergence under half capabilities
-    GMRESSolve<half> gmres_solve_h(A, b, static_cast<half>(u_hlf), n, conv_tol_hlf);
+    // Check convergence under double capabilities
+    GMRESSolve<half> gmres_solve_h(A, b, u_hlf, n, conv_tol_hlf);
 
     gmres_solve_h.solve();
     if (show_plots) { gmres_solve_h.view_relres_plot("log"); }
@@ -118,7 +118,7 @@ TEST_F(GMRESSolveHalfTest, DivergeBeyondHalfCapabilities) {
     EXPECT_LE(gmres_solve_h.get_relres(), 2*conv_tol_hlf);
 
     // Check divergence beyond single capability of the single machine epsilon
-    GMRESSolve<half> gmres_solve_h_to_fail(A, b, static_cast<half>(u_hlf), n, 0.1*u_hlf);
+    GMRESSolve<half> gmres_solve_h_to_fail(A, b, u_hlf, n, 0.1*u_hlf);
     gmres_solve_h_to_fail.solve();
     if (show_plots) { gmres_solve_h_to_fail.view_relres_plot("log"); }
